Use micro task instead of setTimeout and use Uint8Array

+++ javascript/utils/snapshot_memory.js
@@ -1,8 +1,8 @@
 (function () {
   var copy;
   return function () {
-    if (!copy) copy = new Uint32Array(global.Module.wasmMemory.buffer.slice());
-    var view = new Uint32Array(global.Module.wasmMemory.buffer);
+    if (!copy) copy = new Uint8Array(global.Module.wasmMemory.buffer.slice());
+    var view = new Uint8Array(global.Module.wasmMemory.buffer);
     view.fill(0);
     view.set(copy);
   }
